move methods for file writing to FileUtil
